refactor: split priority_select into module with separate test file
Splits the 2214-line priority_select.rs into a cleaner structure:
- priority_select.rs (313 lines) - implementation
- priority_select/tests.rs (1900 lines) - test module

Uses modern Rust module style (named file + directory) rather than
mod.rs pattern. This improves code organization and makes large test
suites easier to navigate without affecting functionality.

All 13 tests pass. No logic changes - pure refactoring.
